// src/dtos/model/get-recipes4-xx-response.ts
/**
 * Recipe
 * Recipe management API.
 *
 * The version of the OpenAPI document: 0.1
 * Contact: kitchen@marmicode.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetRecipes4XXResponse { 
    type?: string;
    title?: string;
    status?: number;
    detail?: string;
    instance?: string;
}